+++ fpga/__init__.py
@@ -1,0 +1,1 @@
+"""FPGA-oriented golden reference package."""